pyproject.toml
[tool.poetry]
name = "api-dynamo"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"
"aws-cdk.core" = "^1.22.0"
"aws-cdk.aws-apigateway" = "^1.22.0"
"aws-cdk.aws-dynamodb" = "^1.22.0"
"aws-cdk.aws-iam" = "^1.22.0"
"aws-cdk.aws-cognito" = "^1.22.0"

[tool.poetry.dev-dependencies]
pylint = "^2.4.4"
mypy = "^0.761"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
